+++ wrapper/src/heap_management.cpp
@@ -24,7 +24,12 @@ extern "C" void *calloc(size_t nmemb, size_t size)
     void *return_value = real_calloc(nmemb, size);
 
      size_t msg_length = message_builder::build_message(
-         libcx.buffer, ELibCall::calloc, libcx.pid, nmemb, size, return_value);
+         libcx.buffer, 
+         ELibCall::calloc, 
+         libcx.pid, 
+         (uint64_t)nmemb, 
+         (uint64_t)size, 
+         (uint64_t)return_value);
 
     libcx.report(msg_length);
